Change the outputs of broadcast_tensors op.
