+++ torch/onnx/symbolic_opset11.py
@@ -9,7 +9,7 @@ import warnings
 import numpy
 
 from torch.onnx.symbolic_helper import parse_args, _unimplemented, _is_tensor_list
-from torch.onnx.symbolic_opset9 import expand, unused, expand_as, mul, zeros_like
+from torch.onnx.symbolic_opset9 import expand, unused, mul
 from torch.nn.modules.utils import _single, _pair, _triple
 from torch.onnx.utils import _add_block, _add_input_to_block, _add_output_to_block
 
@@ -996,19 +996,6 @@ def repeat_interleave(g, self, repeats, dim=None, output_size=None):
     # along the dimension provided, some post-processing is required
     loop_out = g.op("Transpose", loop_out, perm_i=perm_i)
     return reshape(g, loop_out, g.op("Constant", value_t=torch.LongTensor(output_sizes)))
-
-
-def broadcast_tensors(g, self):
-    all_tensors = sym_help._unpack_list(self)
-    t_with_final_shape = zeros_like(g, all_tensors[0])
-
-    # Add operator supports multidirectional broadcasting. So we leverage this function
-    # to infer the final shape generated by the broadcast.
-    for t in all_tensors:
-        t_with_final_shape = add(g, t_with_final_shape, t)
-
-    t_list = [expand_as(g, t, t_with_final_shape) for t in all_tensors]
-    return g.op("SequenceConstruct", *t_list)
 
 
 def normal(g, loc, scale, seed):
